ScenarioDat: fix too many open file descriptors
The importing and writing opened all list of files. This was not
handled well by my system and always gave the error "IOException: too
many file descriptors open".

This is now fixed by opening a file in the queue only when it needs to
be accessed. The importing and writing is still pretty fast.